Let the server choose an XMPP resource for openhab
Otherwhise openHAB's XMPP action may end in an endless loop if the
servers policy on resource conflicts is "drop current". This is openHAB
connects to XMPP server and for some reason the connection from openHAB
to the server is terminated without the server noticing (e.g. your local
router has a short power outtage). Then when openHAB reconnects again
the same XMPP resource is used. The server now sees two XMPP connections
with the same resource and may drop the new one trying to bind the
resource, which causes the connectionClosedOnError(Exception) callback
to get invoked. And the circle starts again.