DSC alarm now sends user code over serial interface
When communicating with a DSC alarm system over the IT-100 interface, a user code is required to disarm the system. This user code was being sent when using an envisalink connection to DSC, and we reuse that code for the serial connection with this change. The code is also padded to 6 digits per the IT-100 Developer Guide.

This fixes #2202