Update..still needs work
Chris Jackson,  I made the changes to the bytes sizes that you
referenced and added the other configurations and association groups. I
dont know how to properly assign all of this but I got most of the grunt
work done. If you can take a look at the manual again and let me know
how to fix the missing parts this will make for a good learning
experience so I can handle these types of configurations on my own.