Added support to Netamato binding for *min_*, *max_*, sum_rain for the intervals 30 minutes, 1 hour, 3 hours, 1 day, 1 week and 1 month. The interval is configurable and defaults to 1 day.
The pull request #2820 that I created didn't have any scaling of converted values, this was giving
the false impression of higher precision than is actually available (many digits of precision).
The values are treated as intermediate vales scaled according to the Significant Figures and
Significant Arithmetic rules which is widely used with scientific data.

Fix issue 1820, if the token was expired, the token was refreshed and a recursive call
was made to re-execute the binding, this would casue null pointer exceptions in the original
execute call to the binding.

Fix issue #3242, handle case when Netatmo provides a null altitude

Added support to also refresh the token if the call to netamoto indicated the token was invalid.

Added a better message when StartCOM CA certificate is missing