MiOS Action - sendMiosAction() API non-functional under openHAB 2 compat
The MiOS Action binding uses a Rule API with an Item as the first
parameter. This breaks under openHAB 2.x/ESH because they won't resolve
to the ESH Item equivalent.

This change moves to declaring the first parameter as Object, and then
dynamically calling getName() on this parameter, which resolves
correctly in both OH 1.x and ESH.

Signed-off-by: Mark Clark <mr.guessed@gmail.com>